move routes to seperate file
