+++ docs/.vitepress/worker/src/index.ts
@@ -1,5 +1,4 @@
 import {
-    Oauth2Routes,
     APIVersion,
     PatreonWebhookTrigger,
     RouteBases,
@@ -10,17 +9,10 @@ import {
     Type,
 } from 'patreon-api.ts'
 
-interface Route {
-    route: string
-    relationship_type: string
-    list?: true
-    requires_id?: true
-}
+import routes from './routes'
 
 export default <ExportedHandler> {
     async fetch () {
-        const id = ':id'
-
         const data = {
             version: APIVersion,
             base: RouteBases.oauth2,
@@ -31,52 +23,7 @@ export default <ExportedHandler> {
                     sha: PATREON_RESPONSE_HEADERS.Sha,
                 }
             },
-            routes: [
-                {
-                    route: Oauth2Routes.campaign(id),
-                    relationship_type: 'campaign',
-                    requires_id: true,
-                },
-                {
-                    route: Oauth2Routes.campaignMembers(id),
-                    relationship_type: Type.Member,
-                    list: true,
-                },
-                {
-                    route: Oauth2Routes.campaignPosts(id),
-                    relationship_type: Type.Post,
-                    list: true,
-                    requires_id: true,
-                },
-                {
-                    route: Oauth2Routes.campaigns(),
-                    relationship_type: Type.Campaign,
-                    list: true,
-                },
-                {
-                    route: Oauth2Routes.identity(),
-                    relationship_type: Type.User,
-                },
-                {
-                    route: Oauth2Routes.member(id),
-                    relationship_type: Type.Member,
-                    requires_id: true,
-                },
-                {
-                    route: Oauth2Routes.post(id),
-                    relationship_type: Type.Post,
-                },
-                {
-                    route: Oauth2Routes.webhook(id),
-                    relationship_type: Type.Webhook,
-                    requires_id: true,
-                },
-                {
-                    route: Oauth2Routes.webhooks(),
-                    relationship_type: Type.Webhook,
-                    list: true,
-                },
-            ] satisfies Route[],
+            routes,
             schemas: {
                 ...Object.keys(SchemaKeys).reduce((obj, key) => ({ ...obj, [key.toLowerCase()]: SchemaKeys[key] }), {}),
                 [Type.PledgeEvent]: SchemaKeys.PledgeEvent,
